Add look behind assertion to avoid including question marks in the end-of-url match.
Comments concerning the fix:

I admit I do not fully understand the reason why the regex does not
match the long sequence of question marks. The matching works up to
about ten question marks and then magically fails. I assume there is
some kind of backtracking mechanism that fails once a certain count is
hit. The middle part of the regex will certainly encompasses the
complete test string at first. When the regex mechanism tries to match
the end-of-url part (see "End with:" in the regex) it will need to
backtrack and reduce the middle match. To me it looks as if this
backtracking has limits. Why the look behind assertion cures this is
also not 100% clear to me.

So this is just a suggestion that needs checking and I won't complain
if it is reverted :)


Concerning breaking the build:

I don't see a reason why a test got added with the explicit intention
of breaking the build. The tested problem does not seem to be any
critical issue that would require immediate fixing. Which is also
suggested by the amount of time it was left unfixed.

I would suggest to add a test marked as incomplete in such cases. And
at the same time a corresponding issue should be opened in our bug
tracker and linked in a comment to the test method.